backend/requirements.txt
# Core FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Optional - InfluxDB support
influxdb-client>=1.38.0

# Authentication (optional - for JWT support)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# SSL/TLS certificates (for MongoDB Atlas)
certifi==2023.11.17

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Async support
httpx>=0.25.0
websockets>=12.0

# Optional monitoring (not required)
# redis>=5.0.0  # Uncomment if using Redis
# prometheus-client>=0.19.0  # Uncomment if using Prometheus
